+++ .gitignore
@@ -13,7 +13,6 @@ node_modules
 # All local .terraform directories
 **/.terraform/*
 
-
 # All .tfstate files
 **/*.tfstate
 **/*.tfstate.*
