+++ package.json
@@ -30,7 +30,6 @@
     "@hmcts/nodejs-healthcheck": "^1.7.0",
     "@hmcts/nodejs-logging": "^3.0.4",
     "@types/config": "^0.0.39",
-    "@types/connect-redis": "^0.0.17",
     "@types/cookie-parser": "^1.4.2",
     "@types/cookies": "^0.7.7",
     "@types/csurf": "^1.11.2",
@@ -41,19 +40,16 @@
     "@types/jsonwebtoken": "^8.5.5",
     "@types/node": "^16.11.0",
     "@types/nunjucks": "^3.2.0",
-    "@types/redis": "^2.8.32",
     "@types/require-directory": "^2.1.2",
     "@types/sass": "^1.16.1",
     "@types/serve-favicon": "^2.5.3",
     "axios": "^0.21.4",
     "ccs-frontend-kit": "^1.1.3",
     "config": "^3.3.6",
-    "connect-redis": "^6.0.0",
     "cookie-encrypter": "^1.0.1",
     "cookie-parser": "^1.4.5",
     "cookies": "^0.8.0",
     "csurf": "^1.11.0",
-    "dotenv": "^10.0.0",
     "express": "^4.17.1",
     "express-nunjucks": "^2.2.5",
     "express-session": "^1.17.2",
@@ -67,7 +63,6 @@
     "moment": "^2.29.1",
     "nodejs-logging": "^1.1.1",
     "nunjucks": "^3.2.3",
-    "redis": "^3.1.2",
     "require-directory": "^2.1.1",
     "sass": "^1.42.1",
     "serve-favicon": "^2.5.0",
@@ -80,7 +75,7 @@
     "@babel/preset-env": "^7.11.5",
     "@types/chai-as-promised": "^7.1.1",
     "@types/copy-webpack-plugin": "^6.0.0",
-    "@types/express-session": "^1.17.0",
+    "@types/express-session": "^1.17.4",
     "@types/jest": "^26.0.13",
     "@types/mini-css-extract-plugin": "^0.9.1",
     "@types/supertest": "^2.0.8",
