+++ src/main/app.ts
@@ -1,3 +1,4 @@
+import { glob } from 'glob';
 const { Express, Logger } = require('@hmcts/nodejs-logging');
 import config = require('config');
 import cookieParser from 'cookie-parser';
@@ -6,42 +7,26 @@ import { Helmet } from './modules/helmet';
 import * as path from 'path';
 import favicon from 'serve-favicon';
 import { Nunjucks } from './modules/nunjucks';
-const { setupDev } = require('./setup/development');
+const { setupDev } = require('./development');
 import  i18next from 'i18next'
 const env = process.env.NODE_ENV || 'development';
 const developmentMode = env === 'development';
-import {NotFoundError} from './errors/errors'
-import fs from 'fs'
+import {HTTPError, NotFoundError} from './errors/errors'
 export const app = express();
-import glob from 'glob'
-import {routeExceptionHandler} from './setup/routeexception'
-
-import {RedisInstanceSetup} from './setup/redis'
-
-
-
-
 app.locals.ENV = env;
 
 // setup logging of HTTP requests
-/**
- * @env Local variables 
- */
- let checkforenvFile = fs.existsSync('.env')
- import {localEnvariables} from './setup/envs'
- if(checkforenvFile){
-   localEnvariables(app);
- }
- 
-
+app.use(Express.accessLogger());
 
 const logger = Logger.getLogger('app');
 
 new Nunjucks(developmentMode, i18next).enableFor(app);
+
+
+
 // secure the application by adding various HTTP headers to its responses
 new Helmet(config.get('security')).enableFor(app);
 
-app.use(Express.accessLogger());
 app.use(favicon(path.join(__dirname, '/public/assets/images/favicon.ico')));
 app.use(express.json())
 app.use(express.urlencoded({extended: false}));
@@ -57,39 +42,55 @@ app.use((req, res, next) => {
 app.enable('trust proxy')
 
 
+
+   
+//Setting up the routes and looping through individuals Paths
+glob.sync(__dirname + '/routes/**/*.+(ts|js)')
+  .map(filename => require(filename))
+  .forEach(route => route.default(app));
+
+ //RFI Related routes 
+ glob.sync(__dirname + '/features/rfi/path.ts')
+ .map(filename => require(filename))
+ .forEach(route => route.default(app));
+
+  //Authentication related routes
+  glob.sync(__dirname + '/features/auth/path.ts')
+  .map(filename => require(filename))
+  .forEach(route => route.default(app));
+
+  //Dashboard related routes
+  glob.sync(__dirname + '/features/dashboard/path.ts')
+  .map(filename => require(filename))
+  .forEach(route => route.default(app));
+
+    //Choose Agreement related routes
+    glob.sync(__dirname + '/features/agreement/path.ts')
+    .map(filename => require(filename))
+    .forEach(route => route.default(app));
+
+  glob.sync(__dirname + '/features/procurement/path.ts')
+  .map(filename => require(filename))
+  .forEach(route => route.default(app));
+
+
+setupDev(app,developmentMode);
+
 /**
- * @RedisClient
- */
-RedisInstanceSetup(app);
-
-
-
-/**
- * @Routable path getting content from default.json
- */
-    let featureRoutes : Array<Object> = config.get('featureDir')
-    featureRoutes?.forEach((aRoute: any) => {
-      glob.sync(__dirname + aRoute?.['path'])
-      .map((filename : string )=> require(filename))
-      .forEach((route: any) => route.default(app));
-    });
-
-/**
- * @developementEnvironment
- *  Setting up development environment
- *  
- */
-    setupDev(app,developmentMode);
-
-/**
- * @ExceptionHandler
  *  All error Handler Routes 
  *  
  */
-    routeExceptionHandler(
-      app,
-      NotFoundError,
-      logger,
-      env
-    )
+ app.use((req, res) => {
+  const notFoundError = new NotFoundError;
+  res.status(notFoundError.statusCode);
+  res.redirect('/404')
+});
 
+
+app.use((err: HTTPError, req: express.Request, res: express.Response) => {
+  logger.error(`${err.stack || err}`);
+  res.locals.message = err.message;
+  res.locals.error = env === 'development' ? err : {};
+  res.status(err.status || 500);
+  res.render('error/500');
+});
