+++ src/main/features/auth/controller/logout.ts
@@ -9,9 +9,5 @@ import * as express from 'express'
  * @param res 
  */
 export const LOGOUT = (req : express.Request, res : express.Response)=> {
-
-  req.session.destroy(function(err) {
-    console.log(err)
-  })
   res.redirect('/');
 }
