+++ src/main/features/rfi/controller/type.ts
@@ -20,23 +20,23 @@ export const GET_TYPE = (req : express.Request, res : express.Response)=> {
  */
 //POST 'rfi/type'
 export const POST_TYPE = (req : express.Request, res : express.Response)=> {
-         let {agreement_id} = req.query;
-         let filtered_body_content_removed_rfi_key =    ObjectModifiers._deleteKeyofEntryinObject(req.body, 'choose_rfi_type');
+   let {agreement_id} = req.query;
+   let filtered_body_content_removed_rfi_key =    ObjectModifiers._deleteKeyofEntryinObject(req.body, 'choose_rfi_type');
 
-         let {ccs_rfi_type} = filtered_body_content_removed_rfi_key;
+   let {ccs_rfi_type} = filtered_body_content_removed_rfi_key;
 
-         switch(ccs_rfi_type){
-            case 'all_online':
-               let redirect_address = `/rfi/online-task-list?agreement_id=${agreement_id}&proc_id=1&event_id=ocds-b5fd17-1`;
-               res.redirect(redirect_address);
-            break;
+   switch(ccs_rfi_type){
+      case 'all_online':
+         let redirect_address = `/rfi/online-task-list?agreement_id=${agreement_id}&proc_id=1&event_id=ocds-b5fd17-1`;
+         res.redirect(redirect_address);
+      break;
 
-            case 'all_offline':
-               const newAddress = RFI_PATHS.GET_UPLOAD_DOC;
-               res.redirect(newAddress);
-            break;
+      case 'all_offline':
+         const newAddress = RFI_PATHS.GET_UPLOAD_DOC;
+         res.redirect(newAddress);
+      break;
 
-            default: res.redirect('/404');
+      default: res.redirect('/404');
    }
   
 }
