--- src/main/public/assets/scripts/session/idle.js
@@ -1,23 +1,0 @@
-var inactiviyTime = 0;
-let logoutlocationURL = window.location.origin + '/oauth/logout'
-
-$(document).ready(function () {
-    // Increment the idle time counter every minute.
-    var idleInterval = setInterval(timerIncrement, 60000); // 1 minute
-
-       // Zero the idle timer on mouse movement.
-    $(this).mousemove(function (e) {
-        inactiviyTime = 0;
-    });
-    $(this).keypress(function (e) {
-        inactiviyTime = 0;
-    });
-});
-
-function timerIncrement() {
-    inactiviyTime = inactiviyTime + 1;
-
-    if (inactiviyTime > 15) { // 2 minutes
-        window.location.href= logoutlocationURL
-    }
-}
