+++ src/main/routes/home.ts
@@ -3,7 +3,5 @@ import {NO_AUTH} from '../common/middlewares/oauthservice/openroutecheck'
 
 export default function(app: Application): void {
   
-  app.get('/', NO_AUTH,  (req, res)=> {
-    res.render('home')
-  });
+  app.get('/', NO_AUTH,  (req, res)=> res.render('home'));
 }
