--- src/main/types/index.ts
@@ -1,9 +1,0 @@
-import session from 'express-session';
-
-declare module 'express-session' {
-    export interface SessionData {
-      user: { [key: string]: any };
-    }
-}
-
-
