+++ src/main/views/webpack/js.njk
@@ -33,6 +33,5 @@
 <script src="/assets/scripts/validations/validate_rfi_upload.js"></script>
 <script src="/assets/scripts/validations/rfi_suppliers.js"></script>
 <script src="/assets/scripts/validations/validate_rfi_address.js"></script>
-<script src="/assets/scripts/session/idle.js"></script>
 
  <script src="/assets/scripts/application.js"></script>
